requirements.txt
crewai>=0.11.0
crewai-tools>=0.0.10
langchain-ibm>=0.0.1
langchain-core>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0